refactor : MVC 패턴을 적용하기 위해 디렉토리 구조 변경
