+++ src/main/java/application/racing/view/InputViewer.java
@@ -1,0 +1,4 @@
+package application.racing.view;
+
+public class InputViewer {
+}
